# memberlist: get rid of exclusive mutex on get worker channel method 