@ortuman 2022-05-25
> Given that memberlist processes incoming messages in single goroutine, I don't think that additional overhead of RW lock is useful here. 

In such a case, and since there's no concurrent access, we could simply do without `workersMu` mutex. 🥳