@pstibrany 2022-05-25
> In such a case, and since there's no concurrent access, we could simply do without `workersMu` mutex. 🥳

That would be nice :) I'm not 100% sure about that however, I see two paths leading to `NotifyMsg` method, and it's not clear if they are from single goroutine or not. I would err on the safe side and keep the lock. However the majority of cases are from single goroutine loop.